Add key, door, and room interactions. Additionally, add a tp cheat command. (#38)

+++ core/src/main/java/io/github/RashRogues/Bomber.java
@@ -9,7 +9,7 @@ public class Bomber extends Enemy {
     private final float BASE_BOMBER_ATTACK_SPEED = 0.1f;
     private final float BASE_BOMBER_MOVE_SPEED = 12f;
 
-    Bomber(Texture texture, int x, int y, float size) {
+    Bomber(Texture texture, float x, float y, float size) {
         super(EntityType.BOMBER, texture, x, y, size);
         this.stats = new EnemyStats(BASE_BOMBER_HEALTH, BASE_BOMBER_DAMAGE, BASE_BOMBER_ATTACK_SPEED, BASE_BOMBER_MOVE_SPEED, this);
     }
